[Refactor] 프로필사진 패키지 분리 #90
