+++ src/main/java/me/snaptime/profile/dto/res/ProfileTagSnapResDto.java
@@ -1,4 +1,4 @@
-package me.snaptime.profile.userProfile.dto.res;
+package me.snaptime.profile.dto.res;
 
 import lombok.Builder;
