+++ src/main/java/me/snaptime/profile/dto/res/UserProfileResDto.java
@@ -1,4 +1,4 @@
-package me.snaptime.profile.userProfile.dto.res;
+package me.snaptime.profile.dto.res;
 
 import lombok.Builder;
 import me.snaptime.user.domain.User;
