+++ src/main/java/me/snaptime/profilePhoto/controller/ProfilePhotoController.java
@@ -1,4 +1,4 @@
-package me.snaptime.profile.profilePhoto.controller;
+package me.snaptime.profilePhoto.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -6,8 +6,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.snaptime.common.CommonResponseDto;
-import me.snaptime.profile.profilePhoto.dto.res.ProfilePhotoResDto;
-import me.snaptime.profile.profilePhoto.service.ProfilePhotoService;
+import me.snaptime.profilePhoto.dto.res.ProfilePhotoResDto;
+import me.snaptime.profilePhoto.service.ProfilePhotoService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
