+++ src/main/java/me/snaptime/user/controller/UserController.java
@@ -8,10 +8,10 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.snaptime.common.CommonResponseDto;
-import me.snaptime.profile.userProfile.dto.res.AlbumSnapResDto;
-import me.snaptime.profile.userProfile.dto.res.ProfileCntResDto;
-import me.snaptime.profile.userProfile.dto.res.ProfileTagSnapResDto;
-import me.snaptime.profile.userProfile.dto.res.UserProfileResDto;
+import me.snaptime.profile.dto.res.AlbumSnapResDto;
+import me.snaptime.profile.dto.res.ProfileCntResDto;
+import me.snaptime.profile.dto.res.ProfileTagSnapResDto;
+import me.snaptime.profile.dto.res.UserProfileResDto;
 import me.snaptime.user.dto.req.SignInReqDto;
 import me.snaptime.user.dto.req.UserReqDto;
 import me.snaptime.user.dto.req.UserUpdateDto;
