+++ src/main/java/me/snaptime/user/repository/UserCustomRepository.java
@@ -1,7 +1,7 @@
 package me.snaptime.user.repository;
 
-import me.snaptime.profile.userProfile.dto.res.AlbumSnapResDto;
-import me.snaptime.profile.userProfile.dto.res.ProfileTagSnapResDto;
+import me.snaptime.profile.dto.res.AlbumSnapResDto;
+import me.snaptime.profile.dto.res.ProfileTagSnapResDto;
 import me.snaptime.user.domain.User;
 
 import java.util.List;
