+++ src/main/java/me/snaptime/user/repository/impl/UserCustomRepositoryImpl.java
@@ -5,8 +5,8 @@ import com.querydsl.core.Tuple;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import me.snaptime.component.url.UrlComponent;
-import me.snaptime.profile.userProfile.dto.res.AlbumSnapResDto;
-import me.snaptime.profile.userProfile.dto.res.ProfileTagSnapResDto;
+import me.snaptime.profile.dto.res.AlbumSnapResDto;
+import me.snaptime.profile.dto.res.ProfileTagSnapResDto;
 import me.snaptime.user.domain.User;
 import me.snaptime.user.repository.UserCustomRepository;
 import org.springframework.stereotype.Repository;
