+++ src/main/java/me/snaptime/user/service/UserService.java
@@ -9,12 +9,12 @@ import me.snaptime.exception.ExceptionCode;
 import me.snaptime.friendShip.dto.res.FriendCntResDto;
 import me.snaptime.friendShip.service.FriendShipService;
 import me.snaptime.jwt.JwtProvider;
-import me.snaptime.profile.profilePhoto.domain.ProfilePhoto;
-import me.snaptime.profile.profilePhoto.repository.ProfilePhotoRepository;
-import me.snaptime.profile.userProfile.dto.res.AlbumSnapResDto;
-import me.snaptime.profile.userProfile.dto.res.ProfileCntResDto;
-import me.snaptime.profile.userProfile.dto.res.ProfileTagSnapResDto;
-import me.snaptime.profile.userProfile.dto.res.UserProfileResDto;
+import me.snaptime.profile.dto.res.AlbumSnapResDto;
+import me.snaptime.profile.dto.res.ProfileCntResDto;
+import me.snaptime.profile.dto.res.ProfileTagSnapResDto;
+import me.snaptime.profile.dto.res.UserProfileResDto;
+import me.snaptime.profilePhoto.domain.ProfilePhoto;
+import me.snaptime.profilePhoto.repository.ProfilePhotoRepository;
 import me.snaptime.snap.repository.SnapRepository;
 import me.snaptime.user.domain.User;
 import me.snaptime.user.dto.req.SignInReqDto;
